# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ms-CRM OpenApi Documentation",
        "description": "ms-CRM OpenApi description",
        "contact": {
            "email": "florian@alt.bzh"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1/",
            "description": "Full ms-CRM API endPoint documentation"
        }
    ],
    "paths": {
...
        "/log/{id}": {
            "get": {
                "tags": [
                    "logAccounting"
                ],
                "summary": "get log by id",
                "description": "Affichage d'une action concernant la facturation",
                "operationId": "4ea08b508c2d1615503fc22fe35a4cd0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "log id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return log Accounting object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LogAccounting"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No log Accounting found"
                    }
                }
            }
        },
        "/log/list": {
            "get": {
                "tags": [
                    "logAccounting"
                ],
                "summary": "Get list of all logs for Accounting",
                "description": "Liste de toutes les actions concernant la facturation : création et modification de factures, de devis, chaque détail, etc ...",
                "operationId": "ca6982e7c6bfffdee64e8c0e6aaff418",
                "responses": {
                    "200": {
                        "description": "return log object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LogAccounting"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "log's list not found"
                    }
                }
            }
        },
        "/log/quote/{id}": {
            "get": {
                "tags": [
                    "logAccounting"
                ],
                "summary": "List of all logs for one quote",
                "description": "Affichage de toutes les actions faites sur un devis",
                "operationId": "8be34dd38cfa3fdc760db3639d7d0f9d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of quote",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return logAccounting object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LogAccounting"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No logAccounting found"
                    }
                }
            }
        },
        "/log/invoice/{id}": {
            "get": {
                "tags": [
                    "logAccounting"
                ],
                "summary": "List of all logs for one invoice",
                "description": "Affichage de toutes les actions faites sur une facture",
                "operationId": "874102d7aece19e85beb68e185e6d1d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of invoice",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return logAccounting object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LogAccounting"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No logAccounting found"
                    }
                }
            }
        },
...
        "/quote/detail/{id}": {
            "get": {
                "tags": [
                    "quoteDetail"
                ],
                "summary": "get quote detail by id",
                "description": "Affichage d'un détail d'un devis",
                "operationId": "42f364e21b60fb7b2b85c6daecb3855a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "quote detail id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return quote detail object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuoteDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No quote detail found"
                    }
                }
            },
            "put": {
                "tags": [
                    "quoteDetail"
                ],
                "summary": "Update detail quote",
                "description": "Mise à jour d'un détail d'un devis et création d'un log",
                "operationId": "da14dff1ceaa38880fe6c8f2e281b6a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "quoteDetail_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuoteDetailUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "return quote detail object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuoteDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Quote detail not updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "quoteDetail"
                ],
                "summary": "Delete quote detail",
                "description": "Suppression d'un détail d'un devis et création d'un log",
                "operationId": "fa21dfa55c8362cf3d6bda4f2db9da19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "quoteDetail_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return quote detail object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuoteDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Quote detail not deleted"
                    }
                }
            }
        },
        "/quote/detail": {
            "post": {
                "tags": [
                    "quoteDetail"
                ],
                "summary": "Store quote detail",
                "description": "Création d'un détail d'un devis et création d'un log",
                "operationId": "946a658bcc976a732765dff5d294c38e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuoteDetailStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "return quote detail object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuoteDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No quote detail store"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "QuoteDetailStoreRequest": {
                "title": "QuoteDetailStoreRequest",
                "description": "Quote detail request body data",
                "required": [
                    "invoice_id",
                    "service_id",
                    "quantity",
                    "excl_tax_price",
                    "tax_id",
                    "tax_percent",
                    "incl_tax_price"
                ],
                "properties": {
                    "quote_id": {
                        "title": "quote_id",
                        "description": "quote's id which this detail is attached",
                        "type": "integer",
                        "example": "4"
                    },
                    "service_id": {
                        "title": "service_id",
                        "description": "id service for this detail, refer to customfield's db",
                        "type": "integer",
                        "example": "8"
                    },
                    "quantity": {
                        "title": "quantity",
                        "description": "quantity of this service",
                        "type": "integer",
                        "example": "2"
                    },
                    "excl_tax_price": {
                        "title": "excl_tax_price",
                        "description": "free tax price",
                        "type": "number",
                        "example": "65.3"
                    },
                    "tax_id": {
                        "title": "tax_id",
                        "description": "tax id for this detail, refer to customfield's db",
                        "type": "integer",
                        "example": "10"
                    },
                    "tax_percent": {
                        "title": "tax_percent",
                        "description": "tax percent for this detail",
                        "type": "number",
                        "example": "102.34"
                    },
                    "incl_tax_price": {
                        "title": "incl_tax_price",
                        "description": "price with tax",
                        "type": "number",
                        "example": "102.37"
                    }
                },
                "type": "object"
            },
            "QuoteDetailUpdateRequest": {
                "title": "QuoteDetailUpdateRequest",
                "description": "Quote detail update request body data",
                "properties": {
                    "service_id": {
                        "title": "service_id",
                        "description": "id service for this detail, refer to customfield's db",
                        "type": "integer",
                        "example": "8"
                    },
                    "quantity": {
                        "title": "quantity",
                        "description": "quantity of this service",
                        "type": "integer",
                        "example": "2"
                    },
                    "excl_tax_price": {
                        "title": "excl_tax_price",
                        "description": "free tax price",
                        "type": "number",
                        "example": "65.3"
                    },
                    "tax_id": {
                        "title": "tax_id",
                        "description": "tax id for this detail, refer to customfield's db",
                        "type": "integer",
                        "example": "10"
                    },
                    "tax_percent": {
                        "title": "tax_percent",
                        "description": "tax percent for this detail",
                        "type": "number",
                        "example": "102.34"
                    },
                    "incl_tax_price": {
                        "title": "incl_tax_price",
                        "description": "price with tax",
                        "type": "number",
                        "example": "102.37"
                    }
                },
                "type": "object"
            },
...
            "LogAccounting": {
                "title": "LogAccounting",
                "description": "LogAccounting model",
                "properties": {
                    "type": {
                        "title": "type",
                        "description": "invoice or quote : 1 pour quote, 2 pour invoice",
                        "type": "integer",
                        "example": "1"
                    },
                    "action": {
                        "title": "action",
                        "description": "action of log, voir AccountingLibs",
                        "type": "integer",
                        "example": "22"
                    },
                    "reference_id": {
                        "title": "reference_id",
                        "description": "reference_id of invoice or quote",
                        "type": "integer",
                        "example": "123456789"
                    },
                    "reference_number": {
                        "title": "reference_number",
                        "description": "reference_number of invoice or quote",
                        "type": "integer",
                        "example": "123456789"
                    },
                    "author_id": {
                        "title": "author_id",
                        "description": "author_id of this log, refer to ??? db",
                        "type": "integer",
                        "example": "2"
                    },
                    "info": {
                        "title": "info",
                        "description": "id du détail ou du paiement concerné",
                        "type": "integer",
                        "example": "18"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LogAccounting"
                }
            },
...
            "QuoteDetail": {
                "title": "QuoteDetail",
                "description": "QuoteDetail model",
                "properties": {
                    "quote_id": {
                        "title": "quote_id",
                        "description": "quote's id which this detail is attached",
                        "type": "integer",
                        "example": "4"
                    },
                    "service_id": {
                        "title": "service_id",
                        "description": "id service for this detail, refer to customfield's db",
                        "type": "integer",
                        "example": "8"
                    },
                    "quantity": {
                        "title": "quantity",
                        "description": "quantity of this service",
                        "type": "integer",
                        "example": "2"
                    },
                    "excl_tax_price": {
                        "title": "excl_tax_price",
                        "description": "free tax price",
                        "type": "number",
                        "example": "65.3"
                    },
                    "tax_id": {
                        "title": "tax_id",
                        "description": "tax id for this detail, refer to customfield's db",
                        "type": "integer",
                        "example": "10"
                    },
                    "tax_percent": {
                        "title": "tax_percent",
                        "description": "tax percent for this detail",
                        "type": "number",
                        "example": "102.34"
                    },
                    "incl_tax_price": {
                        "title": "incl_tax_price",
                        "description": "price with tax",
                        "type": "number",
                        "example": "102.37"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "QuoteDetail"
                }
            },
...
        }
    },
...
    "security": [
        []
    ]
}